cleanup custom theme example
