--- examples/custom_theme_example/lib/main.dart
@@ -1,40 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'themes/app_theme.dart';
-
-void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return AppTheme(
-      data: themeDataBlue,
-      child: const MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: AppTheme.of(context).color,
-      body: const Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: const <Widget>[
-            Text(
-              'This is the home page',
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
