+++ examples/custom_theme_example/pubspec.yaml
@@ -9,11 +9,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  widgetbook_annotation: ^3.0.0-rc.1
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
   widgetbook: ^3.0.0-rc.1
 
 flutter:
